GH-920: Improvements in ack / DLPR. Updated docs.
